Move symlink to summarize
